Draw nodes on top in neural network example
The example increases the thickness of one of the node borders,
which means that with lines drawn on top, the black lines reach
into the red border. If we instead draw nodes on top, the lines
look as though they only go up to the edge of the thickened border.